before release,beta 1.01

+++ AzCare-APIServices/src/main/java/com/project/AzCar/Repositories/Coupon/UserCouponRepository.java
@@ -6,11 +6,13 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import com.project.AzCar.Entities.Coupon.UserCoupon;
+
 @Repository
-public interface UserCouponRepository  extends JpaRepository<UserCoupon, Integer> {
-	
-	@Query(value = "select * from user_coupons r where r.user_id = :id",nativeQuery = true)
+public interface UserCouponRepository extends JpaRepository<UserCoupon, Long> {
+
+	@Query(value = "select * from user_coupons r where r.user_id = :id", nativeQuery = true)
 	UserCoupon findUserCouponByUserId(@Param("id") int id);
-	@Query(value = "select * from user_coupons r where r.coupon_id = :id",nativeQuery = true)
+
+	@Query(value = "select * from user_coupons r where r.coupon_id = :id", nativeQuery = true)
 	UserCoupon findUserCouponByCouponId(@Param("id") int id);
 }
